requirements.txt
asgiref==3.3.2
Django==3.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt
django-mssql-backend==2.8.1
drf_yasg